pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aitf"
version = "0.1.0"
description = "AI Test Framework — test data management, stub orchestration, and CI tooling"
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0",
    "SQLAlchemy>=2.0",
    "paramiko>=3.0",
    "Flask>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
aitf = "aitf.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["aitf*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
